[refactor] Search route 폴더 구조 변경, Card 컴포넌트 props 변경
- 마이페이지 테스트에도 북마크 추가/삭제 기능 추가
