+++ src/view/Feed.js
@@ -61,17 +61,7 @@ const Feed = () => {
             getMoreDatas={getMoreDatas}
           >
             {testsByTag.map((test) => (
-              <Card
-                key={`test ${test.uid}`}
-                uid={test.uid}
-                title={test.title}
-                coverImg={test.coverImg}
-                makerName={test.maker.name}
-                makerProfile={test.maker.profileImg}
-                sharedCnt={test.sharedCnt}
-                participatedCnt={test.participantsCnt}
-                testLink={test.testLink}
-              />
+              <Card key={`test ${test.uid}`} test={test} maker={test.maker} />
             ))}
           </InfinScroll>
         )}
